fourmc: rename planner helper
- Replace FourMcPlanning with FourMcPlanner companion helpers.

- Remove old FourMcParallelPlanner entry point.

- Update scans/tables to pass planner and reuse helpers.
